Increase the frequency to find replacement counts
The index "idx_count" gives a great performance so the counts are usually retrieved in about 5 s. Then we reduce the time for cache refresh to 5 min.

On the other hand, this index gives poor performance on local H2 database, so in this case we keep to old value to 10 min.