Remove replacement list cache
Simplify the replacement list behaviour. Only load/refresh the replacement list when entering in the page. In any change is done in a different tab, no modification will be displayed. The only supported case is when reviewing a complete subtype, which results in a count zero for that subtype, which is displayed as it happens in the same tab.